requirements.txt
#core
torch>=2.2.0
torchvision
timm
numpy
pandas
scikit-learn
opencv-python
matplotlib
Pillow
PyYAML

#explainibility
captum
shap
scikit-image

#wsi
openslide-python

#ui
tk

#optional
tensorboard
scipy
seaborn
